backend/requirements.txt
fastapi==0.115.5
uvicorn[standard]==0.32.1
numpy==2.1.3
python-multipart==0.0.17
pyyaml==6.0.2
pydantic==2.10.3
